Fix Redis expire type mismatch and update dependencies
- Fix E0308 error: Change window_secs cast from i64 to usize in middleware.rs
  to match redis crate's expire() method signature
- Remove unused opentelemetry dependencies from shared crate
- Update tower to 0.5
- Fix CORS configuration in all services (unwrap_or to unwrap_or_else)

+++ suvidha-one-backend/shared/src/middleware.rs
@@ -125,7 +125,7 @@ pub async fn rate_limit_middleware(
     })?;
     
     // Set expiry on the key
-    let _: () = conn.expire(&rate_key, window_secs as i64).await.map_err(|e| {
+    let _: () = conn.expire(&rate_key, window_secs as usize).await.map_err(|e| {
         tracing::error!("Redis expire error: {}", e);
         AppError::Cache(format!("Rate limit expire failed: {}", e))
     })?;
